% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating_data.R
\name{add_metadata}
\alias{add_metadata}
\title{Adding metadata to raw_data}
\usage{
add_metadata(.data, .metadata)
}
\arguments{
\item{.data}{A tibble}

\item{.metadata}{A tibble, metadata of the file (s)}
}
\description{
Adding metadata to raw_data
}
\seealso{
Other add functions: \code{\link{add_population}}
}
\concept{add functions}
