% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating_data.R
\name{add_population}
\alias{add_population}
\title{Add the population Data to the final data}
\usage{
add_population(.data, pop_data = borno_pop)
}
\arguments{
\item{.data}{A tibble}

\item{pop_data}{A tibble with two columns (\code{location} and \code{population}). See the
\code{borno_pop} format.}
}
\value{
A tibble
}
\description{
Add the population Data to the final data
}
\seealso{
Other add functions: \code{\link{add_metadata}}
}
\concept{add functions}
