% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{borno_pop}
\alias{borno_pop}
\title{Population of Borno, 2016 update from OCHA}
\format{A data.frame with two columns:
\itemize{
\item \code{location} representing lgas of Borno
\item \code{population} representing population of each lga in 2016, estimated using the
2006 census data.
}}
\usage{
borno_pop
}
\description{
Population of Borno, 2016 update from OCHA
}
\keyword{datasets}
