% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating_data.R
\name{check_excel}
\alias{check_excel}
\title{A little function to check if the file has an excel extension}
\usage{
check_excel(.file_path)
}
\arguments{
\item{.file_path}{the path to the excel file}
}
\description{
A little function to check if the file has an excel extension
}
