% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating_data.R
\name{check_pdfs}
\alias{check_pdfs}
\title{A little function to check if a file is a pdf and return a stop otherwise}
\usage{
check_pdfs(.file_path)
}
\arguments{
\item{.file_path}{the path to the pdf file}
}
\description{
A little function to check if a file is a pdf and return a stop otherwise
}
