% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating_data.R
\name{convert_file}
\alias{convert_file}
\title{Converting one PDF text into tibble format}
\usage{
convert_file(.file_path, is_excel = FALSE)
}
\arguments{
\item{.file_path}{A character vector of length one. The path to a pdf file}

\item{is_excel}{A logical. Testing if we should update from excel or pdf file}
}
\value{
A tibble
}
\description{
Converting one PDF text into tibble format
}
\seealso{
Other Extracting functions: \code{\link{convert_files}},
  \code{\link{create_metadata}}
}
\concept{Extracting functions}
