% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating_data.R
\name{convert_files}
\alias{convert_files}
\title{Converting all pdfs withing a folder to a data.frame}
\usage{
convert_files(.folder_path, keepwide = FALSE)
}
\arguments{
\item{.folder_path}{A character vector of length one, path to the folder of
files.}

\item{keepwide}{Logical. Should we keep final wide or long format?}
}
\value{
A tibble
}
\description{
Converting all pdfs withing a folder to a data.frame
}
\seealso{
Other Extracting functions: \code{\link{convert_file}},
  \code{\link{create_metadata}}
}
\concept{Extracting functions}
