% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating_data.R
\name{convert_to_long}
\alias{convert_to_long}
\title{Converting data to long format}
\usage{
convert_to_long(.data)
}
\arguments{
\item{.data}{A tibble}
}
\value{
A tibble
}
\description{
Converting data to long format
}
\concept{Extracting Functions}
