% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating_data.R
\name{file_update}
\alias{file_update}
\title{Update historic data using one file}
\usage{
file_update(.file_path, keepwide = FALSE, update = TRUE)
}
\arguments{
\item{.file_path}{A character vector of length One, path to one pdf file}

\item{keepwide}{Logical Should we keep final long or wide format?}

\item{update}{Logical. Update with historical data or not?}
}
\value{
A tibble with updated data
}
\description{
Update historic data using one file
}
\seealso{
Other update functions: \code{\link{folder_update}}
}
\concept{update functions}
