% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historic_long_data}
\alias{historic_long_data}
\title{Historic data in long and wide format}
\format{A Tible. Representing historic extracted data, in wide or long format}
\usage{
historic_long_data
}
\description{
Historic data in long and wide format
}
\keyword{datasets}
