% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historic_wide_data}
\alias{historic_wide_data}
\title{Historic data in long and wide format}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3950 rows and 23 columns.}
\usage{
historic_wide_data
}
\description{
Historic data in long and wide format
}
\keyword{datasets}
