% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating_data.R
\name{create_metadata}
\alias{create_metadata}
\title{Extracting the metadata from one pdf file.}
\usage{
create_metadata(.file_path, is_excel = FALSE)
}
\arguments{
\item{.file_path}{A character vector of length one, indicating the path to a pdf file}

\item{is_excel}{A logical, checking whether metadata should be
extracted from an excel file or not. Default to false}
}
\value{
A tibble
}
\description{
Extracting the metadata from one pdf file.
}
\seealso{
Other Extracting functions: \code{\link{convert_files}},
  \code{\link{convert_file}}
}
\concept{Extracting functions}
